refactor(tests): move all unit tests to a crate-internal src/tests/ tree
Per the standing convention, unit tests live in src/tests/*.rs wired via
#[cfg(test)] mod tests; in lib.rs, not inline in the source files. Each source
module keeps only production code.

- one src/tests/<module>.rs per source module that had tests; glob test imports
  expanded to explicit absolute crate:: paths.
- shared fixtures (SuiteStats::available, DiffCounter::counted) centralized in
  src/tests/mod.rs, in scope for every test submodule; per-suite fixture builders
  (Test/SummarySuite/Summary, Benchmark::insert_test) live with their tests.
- golden fixtures moved to src/tests/fixtures/; assert_matches_fixture path and
  .gitattributes pin updated.
- Xlsx::worksheets / Xlsx::sheet are now pub, reachable from the relocated test.
