+++ solx-benchmark-converter/src/output/summary/diff_counter.rs
@@ -64,33 +64,4 @@ impl DiffCounter {
         };
         format!("⚠️ {} of {}{delta}", commas(self.diffs), commas(self.cells))
     }
-
-    /// A counter with the given tallies, for the output-verdict tests.
-    #[cfg(test)]
-    pub fn counted(cells: u64, diffs: u64, delta: i128) -> Self {
-        Self {
-            cells,
-            diffs,
-            delta,
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::output::summary::diff_counter::DiffCounter;
-
-    #[test]
-    fn diff_counter_skips_uncollected_pairs_and_sums_deltas() {
-        let mut counter = DiffCounter::default();
-        assert!(!counter.observe(0, 0));
-        assert!(!counter.collected());
-        assert!(!counter.observe(100, 100));
-        assert!(counter.observe(90, 100));
-        assert!(counter.observe(115, 100));
-        assert!(counter.collected());
-        assert_eq!(counter.cells, 3);
-        assert_eq!(counter.diffs, 2);
-        assert_eq!(counter.delta, 5);
-    }
 }
