+++ solx-benchmark-converter/src/tests/fixtures/degraded-harness.md
